Use fixed sizes all over the place.

+++ include/dnet/interface.h
@@ -161,7 +161,7 @@ int dnet_write_object(struct dnet_node *n, struct dnet_io_control *ctl, void *re
  * Returns negative error value in case of error.
  */
 int dnet_write_file(struct dnet_node *n, char *file, unsigned char *id,
-		off_t offset, size_t size, unsigned int aflags);
+		uint64_t offset, uint64_t size, unsigned int aflags);
 
 /*
  * Log flags.
@@ -454,14 +454,14 @@ void *dnet_req_header(struct dnet_data_req *r);
 void *dnet_req_data(struct dnet_data_req *r);
 void *dnet_req_private(struct dnet_data_req *r);
 
-void dnet_req_set_header(struct dnet_data_req *r, void *header, size_t hsize, int free);
-void dnet_req_set_data(struct dnet_data_req *r, void *data, size_t size, off_t offset, int free);
-void dnet_req_set_fd(struct dnet_data_req *r, int fd, off_t offset, size_t size, int close);
+void dnet_req_set_header(struct dnet_data_req *r, void *header, uint64_t hsize, int free);
+void dnet_req_set_data(struct dnet_data_req *r, void *data, uint64_t size, uint64_t offset, int free);
+void dnet_req_set_fd(struct dnet_data_req *r, int fd, uint64_t offset, uint64_t size, int close);
 void dnet_req_set_flags(struct dnet_data_req *r, unsigned int mask, unsigned int flags);
 void dnet_req_set_complete(struct dnet_data_req *r,
 		void (* complete)(struct dnet_data_req *r), void *priv);
 
-struct dnet_data_req *dnet_req_alloc(struct dnet_net_state *st, size_t hsize);
+struct dnet_data_req *dnet_req_alloc(struct dnet_net_state *st, uint64_t hsize);
 void dnet_req_destroy(struct dnet_data_req *r);
 
 int dnet_data_ready(struct dnet_net_state *st, struct dnet_data_req *r);
